@@ -9,72 +9,68 @@
     "analyze": "cross-env NODE_ENV=production analyze=yes webpack --progress --hide-modules"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "1.2.19",
-    "@fortawesome/free-regular-svg-icons": "5.9.0",
-    "@fortawesome/free-solid-svg-icons": "5.9.0",
-    "@fortawesome/react-fontawesome": "0.1.4",
+    "@fortawesome/fontawesome-svg-core": "1.2.25",
+    "@fortawesome/free-regular-svg-icons": "5.11.2",
+    "@fortawesome/free-solid-svg-icons": "5.11.2",
+    "@fortawesome/react-fontawesome": "0.1.7",
     "bootstrap": "4.3.1",
     "classnames": "2.2.6",
-    "core-js": "3.1.4",
-    "date-fns": "2.0.0-beta.3",
-    "faker": "4.1.0",
-    "lodash": "4.17.15",
-    "query-string": "6.8.1",
-    "react": "16.8.6",
-    "react-datepicker": "^2.9.6",
-    "react-dom": "16.8.6",
-    "react-redux": "7.1.0",
-    "react-router-dom": "5.0.1",
-    "react-text-mask": "^5.4.3",
+    "core-js": "3.4.1",
+    "date-fns": "2.7.0",
+    "query-string": "6.8.3",
+    "react": "16.11.0",
+    "react-dom": "16.11.0",
+    "react-redux": "7.1.3",
+    "react-router-dom": "5.1.2",
     "redux": "4.0.4",
-    "redux-observable": "1.1.0",
-    "rxjs": "6.5.2",
-    "yup": "0.27.0"
+    "yup": "0.27.0",
+    "lodash": "4.17.15",
+    "redux-observable": "1.2.0",
+    "rxjs": "6.5.3",
+    "react-text-mask": "5.4.3",
+    "react-datepicker": "2.10.0"
   },
   "devDependencies": {
-    "@types/faker": "4.1.5",
-    "@types/jest": "24.0.15",
-    "@types/node": "12.6.8",
-    "@types/react": "16.8.23",
-    "@types/react-dom": "16.8.5",
-    "@types/react-router-dom": "4.3.4",
-    "@types/react-redux": "7.1.1",
+    "@types/node": "12.12.7",
+    "@types/react": "16.9.11",
+    "@types/react-dom": "16.9.4",
+    "@types/react-router-dom": "5.1.2",
+    "@types/react-redux": "7.1.5",
     "@types/classnames": "2.2.9",
-    "@types/yup": "0.26.22",
-    "@types/lodash": "4.14.136",
-    "@types/react-datepicker": "^2.9.2",
-    "@types/react-text-mask": "^5.4.6",
-    "cross-env": "5.2.0",
-    "typescript": "3.5.3",
-    "webpack": "4.37.0",
-    "webpack-cli": "3.3.6",
-    "eslint-loader": "2.2.1",
-    "eslint": "6.1.0",
+    "@types/yup": "0.26.24",
+    "@types/lodash": "4.14.146",
+    "@types/react-datepicker": "2.9.4",
+    "@types/react-text-mask": "5.4.6",
+    "cross-env": "6.0.3",
+    "typescript": "3.7.2",
+    "webpack": "4.41.2",
+    "webpack-cli": "3.3.10",
+    "eslint-loader": "3.0.2",
+    "eslint": "6.6.0",
     "eslint-import-resolver-webpack": "0.11.1",
-    "@typescript-eslint/parser": "1.13.0",
-    "@typescript-eslint/eslint-plugin": "1.13.0",
+    "@typescript-eslint/parser": "2.7.0",
+    "@typescript-eslint/eslint-plugin": "2.7.0",
     "eslint-plugin-import": "2.18.2",
     "eslint-plugin-promise": "4.2.1",
-    "eslint-plugin-react": "7.14.3",
-    "eslint-plugin-react-hooks": "1.6.1",
+    "eslint-plugin-react": "7.16.0",
+    "eslint-plugin-react-hooks": "2.2.0",
     "eslint-plugin-jsx-a11y": "6.2.3",
-    "eslint-plugin-prettier": "3.1.0",
-    "eslint-config-prettier": "6.0.0",
-    "prettier": "1.18.2",
-    "ts-loader": "6.0.4",
-    "sass-loader": "7.1.0",
-    "node-sass": "4.12.0",
-    "css-loader": "3.1.0",
-    "style-loader": "0.23.1",
+    "eslint-plugin-prettier": "3.1.1",
+    "eslint-config-prettier": "6.5.0",
+    "prettier": "1.19.1",
+    "ts-loader": "6.2.1",
+    "sass-loader": "8.0.0",
+    "node-sass": "4.13.0",
+    "css-loader": "3.2.0",
+    "style-loader": "1.0.0",
     "postcss-loader": "3.0.0",
-    "autoprefixer": "9.6.1",
-    "file-loader": "4.1.0",
-    "url-loader": "2.1.0",
+    "autoprefixer": "9.7.1",
+    "file-loader": "4.2.0",
+    "url-loader": "2.2.0",
     "mini-css-extract-plugin": "0.8.0",
     "optimize-css-assets-webpack-plugin": "5.0.3",
-    "uglifyjs-webpack-plugin": "2.1.3",
-    "clean-webpack-plugin": "3.0.0",
-    "webpack-bundle-analyzer": "3.3.2",
+    "uglifyjs-webpack-plugin": "2.2.0",
+    "webpack-bundle-analyzer": "3.6.0",
     "html-webpack-plugin": "3.2.0"
   }
 }
