@@ -2,14 +2,16 @@ import { ActionType } from 'data/actionTypes';
 
 import { AppAction, Dictionary } from './baseTypes';
 
-export function createAction<T>(type: ActionType): (data?: T, mod?: string) => AppAction<T> {
-  return (data: any, mod?: string) => ({ type, data, mod });
+export function createAction<T>(
+  type: ActionType
+): (data?: T, mod?: string, callBack?: (error?: any) => void) => AppAction<T> {
+  return (data: any, mod?: string, callBack?: (error?: any) => void) => ({ type, data, mod, callBack });
 }
 
 export function createReducer<T extends { [index: string]: any }>(
   initialState: () => T,
   info: Dictionary<string | ((state: T, action: AppAction<any>) => T)>,
-  supportReset: boolean = true
+  supportReset = true
 ) {
   return (state: T = initialState(), action: AppAction<any>): T => {
     if (supportReset && action.type === ActionType.CORE_RESET_STATE) return initialState();
