@@ -8,9 +8,18 @@ import './dropdownButton.scss';
 interface DropdownButtonProps extends ButtonProps {
   label?: React.ReactElement;
   rightAlignment?: boolean;
+  groupAppend?: boolean;
+  closeAfterClick?: boolean;
 }
 
-export const DropdownButton: React.FC<DropdownButtonProps> = ({ label, rightAlignment, children, ...other }) => {
+export const DropdownButton: React.FC<DropdownButtonProps> = ({
+  label,
+  rightAlignment,
+  children,
+  groupAppend,
+  closeAfterClick,
+  ...other
+}) => {
   const [show, setShow] = useState(false);
   const focusHolder = useRef<any>(null);
   const activate = (e: React.MouseEvent) => {
@@ -22,17 +31,21 @@ export const DropdownButton: React.FC<DropdownButtonProps> = ({ label, rightAlig
   const deactivate = () => setShow(false);
   return (
     <div
-      className="rm-dropdownButton btn-group"
+      className={classNames('rm-dropdownButton btn-group', { 'input-group-append': groupAppend })}
       onMouseDown={activate}
       onBlur={deactivate}
       tabIndex={0}
       role="button"
       ref={focusHolder}>
-      <Button className="dropdown-toggle" onClick={activate} {...other}>
+      <Button className="dropdown-toggle" {...other}>
         {label}
       </Button>
       {show && (
-        <div className={classNames('dropdown-menu show', { 'dropdown-menu-right': rightAlignment })}>{children}</div>
+        <div
+          onClick={closeAfterClick ? deactivate : undefined}
+          className={classNames('dropdown-menu show', { 'dropdown-menu-right': rightAlignment })}>
+          {children}
+        </div>
       )}
     </div>
   );
